+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 (Relative to [`bossy` v0.2.1](https://github.com/BrainiumLLC/bossy/blob/master/CHANGELOG.md#021-2021-01-08))
 
 - Added methods to `diva::Command` for setting the current working directory.
+- Fixed Windows build.